requirements.txt
openbabel-wheel>=3.1.1
tqdm>=4.66.0
